Update to latest MicroPython core, v1.5.2-76.
There are lots of changes, but most of them are to support configuration
of more sophisticated object models, such as 64-bit NaN boxing (not used
on the microbit).

As usual there are bug fixes, and a few new features.  One such feature
is that dir() now lists instance members of an object.