microbit: Add acceleromter set_range method.
set_range is available in MakeCode and has been requested in MicroPython to
be able to create equivalent programmes and lessons.

get_range has not been requested, is not available in MakeCode, and there
is not a lot of cases where it would be useful, so it's not included here.
Getting/setting the sampling rate is similarly not exposed in MakeCode and
has never been requested yet in MakeCode nor MicroPython, so is also not
included.