Add MicroBitDisplay.on(), .off(), and .is_on().
Turning off the display will preserve the state of the display (pausing any
animation).  Turning on the display will restore the state (resuming any
animation, not restarting).

is_on() will return True if the display is on and False if the display is
off.

Also, add drive-by documentation for the display driver, especially of the PWM
mechanism.