microbitmusic: factor out music stop code.
Move the code that stops the current music that is playing into its own
function. By moving to a common function for disabling music, this also
fixes an issue where we would previously free the wrong pin when calling
music.play().

NOTE: music.pitch is still broken, and will be fixed in a subsequent
commit.

NOTE: If someone ran python code in an interrupt handler to start more
music, it is still possible to get the global variables into an
inconsistent state.