microbit: Consolidate all DAL components into one place.
The declarations are in microbitdal.h and object instances in main.cpp.

Keeping them together makes it easier to track what components of the DAL
are used.  And these instances are C++ objects with non-trivial
constructors so it makes sense to list them in the file that does all the
initialisation.

The only component that is not in main.cpp is MicroBitSerial which is in
mphalport.cpp, because it is private to that file.