microbit: Implement radio.receive_full(), returns data, rssi, ticks_us.
The rssi is in dBm and ticks_us is the microsecond timestamp with 30-bit
resolution.  Movitation for having the timestamp in us is:

- A millisecond is a long time in the land of radio frequencies.

- A micro:bit sending raw packets as fast as it can will send them with
  about 700us between packets, so sub-ms resolution is required to
  distinguish such packets.

- It's good sense to support other radio/nRF devices that may send at
  faster rates than a micro:bit.

- time.ticks_us() and time.ticks_diff() are now provided on the micro:bit
  so these functions can be used to deal with microsecond timestamps.

With microseconds it makes sense to use large numbers (ie 30-bit) so that
a period of greater than 1 second can be encoded.  4 bytes for the
timestamp only adds 16 bytes to the default buffer size for the radio
(with a queue length of 3, plus 1 tx buffer) so that is acceptable.