microbit: Consolidate microbit module definitions into modmicrobit.h.
This patch moves all the definitions of types and functions for things in
the microbit module into modmicrobit.h, removing all of the individual
header files, and reorganising modmicrobit.h into sections for each type.
Except for microbitimage.h which cannot be merged because it contains C++
specific code, and is instead renamed to microbit_image.h.

In addition to tidying things up, this should now fix issues when building
on systems with case-insensitive filenames, because there are no longer any
filenames which clash with DAL filenames (eg microbitpin.h vs
MicroBitPin.h).