microbit: Modify PWM driver so it uses static RAM, not heap RAM.
Prior to this patch the PWM driver would allocate RAM on the heap each time
the duty cycle was changed on a pin, or the PWM period was changed.  This
behaviour can lead to a highly fragmented heap and frequent GC pauses when
the duty cycle and/or period are changed often in a loop, which is quite
common in programs that want to vary the PWM output.

This patch changes this behaviour so the PWM driver uses statically
allocated RAM.  It uses 96 bytes of RAM and supports PWM on up to 19 pins
at once, which is as many as the device has.