microbit: Change audio module to use TIMER2.
TIMER1 is used by low-level mbed timing routines so cannot be used by the
audio module.  In particular the functions microbit.sleep() and
microbit.running_time() both rely on TIMER1 and do not work if the audio
module uses TIMER1.

TIMER0 is used by the PWM driver, but TIMER2 is available to use for audio.