microbit: Allow audio module to reinitialise after a soft reset.
Since audio_init() allocates memory on the heap it must re-run after each
soft reset.  This patch makes such a thing work by removing the local
static variable "initialised" and using instead the value of audio_buffer
to check if the audio is initialised or not.  This value is reset upon a
soft reset.