microbit: Change radio to use a fixed RX buffer to prevent RX errors.
Prior to this patch the RX buffer for the radio would move to always point
to the end of the RX queue, and the hardware would fill in any incoming
radio data into this movable location.  This RX destination would increase
and decrease in RAM when packets were added and removed.  The problem with
this approach is that it would lead to race conditions where the hardware
was writing to the buffer while the software was decreasing the buffer
after popping a packet off the queue.

This patch changes the behaviour so that incoming RX packets are always
stored into the same, fixed location.  This location is the beginning of
the buffer where the TX packets also go out from.  It's safe to use the
same buffer for TX and RX because they can only happen independently, never
at the same time.  Once a valid packet is received it is copied to the end
of the RX queue, but only if there is enough room.  This incurs a small
additional overhead to do the copy but this shouldn't take much time
because packets are small.

As a bonus, it's now possible to store more packets in the RX queue for the
same amount of allocated RAM because there no longer needs to be a whole
packet's worth of room available in the queue for the next incoming packet.