microbit: Consolidate neopixel support code into neopixel module.
This patch puts all the neopixel code directly into the neopixel module:
getting, setting and clearing pixels is done directly in the class methods,
and the only external code is the sendNeopixelBuffer() function which does
the low-level output of the data.

This saves code size, makes it a little more efficient, eliminates the use
of malloc to allocate the memory for the pixel data buffer (it now uses the
uPy heap not the mbed heap and so doesn't leak memory), and allows in the
future to support different colour layouts/formats of the pixels by making
the RGB index macros dynamic variables.