Add "active" to increase specificity
There are active piezo buzzers that include their own oscillator to produce a tone of a predetermined frequency when a current is applied and passive piezo buzzers that require an alternating current to produce a tone, that tone being of the frequency at which the current alternates.  This section and the next fail to recognize that difference.  A passive buzzer used in these exercises would at best produce a click when a current was switched on or off.
I suppose one could potentially chalk it up to an issue of semantics, as some might call an active buzzer a buzzer and a passive buzzer a speaker, but I think it is fairly common to refer to a low fidelity piezo as a buzzer.