docs: Embellish explanation of piezo buzzer types.
Expand the introduction to piezo buzzers to explain that there are two
types with slightly different uses.  It is important that the student/user
be using an active buzzer for the example code to work.

Also add text stating that a passive piezo buzzer could be used as an
alternative to a speaker and updated the warning to say that active piezo
buzzers shouldn't be used.