microbit: Tune parser memory allocation policy.
Combined with the previous change to rule_stack_t, this makes sure that
the number of bytes allocated for the parser's rule stack is unchanged for
the initial allocation, and for subsequent resizes of the stack.

What does change is that now the rules stack can hold more entries for the
same number of bytes.  The result is that there will be less resizes of the
rule stack.