docs: update neopixel page for 5V/3V clarity (#596)
This change updates the images and text in the neopixel docs page
to ensure that we're clear the micro:bit cannot be powered with 5V.

The SVG is included as 'source' for the image rendered (with Inkscape)
to simplify further updates. It's not used in the docs directly to
ensure that the docs continue to render well across all browsers,
including some of the very old versions we see in schools.