Makefile: Use yotta command instead of yt.
When yotta is installed via the apt package manager on Debian/Ubuntu, the
`yt` shorthand alternative for the yotta command is unavailable due to a
clash with an identically-named command from the python3-yt package.

This update ensures the default yotta command is used when building
MicroPython via the Makefile.