microbit: Remove accelerometer set/get_rate and get_range.
To save memory we need to keep the additions to a minimum. The
sampling rate is not exposed in MakeCode and has never been
requested yet in MakeCode nor MicroPython.

set_range is available in MakeCode and has been requested in
MicroPython to be able to create equivalent programmes and lessons.
get_range has not been requested, is not available in MakeCode,
and there is not a lot of cases where it would be useful, so it's
been also removed to save memory.