docs: Remove radio.RATE_250KBIT as it has been dropped from V2.
The constant is still present in the micro:bit V1 MicroPython
port, but it is not in the V2 MicroPython port.
The value from the constant (`2`) can be used as an argument in
the V2 MicroPython `radio.config(data_rate=2)`, but because it is
not officially supported in the V2 nRF52833 microcontroller we
cannot guarantee it will always work, so the constant is
deprecated in V1 and removed in V2.