docs: Update accelerometer page with better range info.
Correct old notes about not being able to change the range.
Remove note already mentioned in the header, and move to header
note about needing to call an accelerometer function regulartly.

Indicate in the header the default range so that it can be removed
from each function description.

Use :param: and :return: in the function descriptions.